Rename application
Change original name from goreact to hlsadmin. Also updated the README.
