+++ cmd/hlsadmin/cli/noui.go
@@ -16,7 +16,7 @@ package cli
 
 import (
 	"fmt"
-	"goweb/internal/server"
+	"hlsdevkit/orch/internal/server"
 	"log"
 	"net/http"
