+++ cmd/hlsadmin/cli/noui_test.go
