+++ cmd/hlsadmin/cli/start_test.go
