+++ cmd/hlsadmin/cli/ui_test.go
