+++ internal/usermgmt/authuser_test.go
@@ -18,7 +18,7 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"goweb/internal"
+	"hlsdevkit/orch/internal"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
